using routes.tsx instead of registery.json
