+++ common/v2/features/CreateWallet/index.ts
@@ -1,2 +1,3 @@
-export { default } from './CreateWallet';
-export { default as createWalletRoutes } from './routes';
+export { default as CreateWallet } from './CreateWallet';
+export { Mnemonic } from './Mnemonic';
+export { Keystore } from './Keystore';
