+++ common/v2/features/Dashboard/Settings/index.ts
@@ -1,5 +1,3 @@
-export { default as settingsRoutes } from './routes';
-export * from './components';
-export { default as SettingsFlow } from './Settings';
+export { default as Settings } from './Settings';
 export { Import } from './Import';
 export { Export } from './Export';
