--- common/v2/features/Dashboard/Settings/routes.ts
@@ -1,24 +1,0 @@
-import Settings from './Settings';
-import { Export } from './Export';
-import { Import } from './Import';
-
-export default [
-  {
-    name: 'Settings',
-    path: '/settings',
-    exact: true,
-    component: Settings
-  },
-  {
-    name: 'Import',
-    path: '/settings/import',
-    exact: true,
-    component: Import
-  },
-  {
-    name: 'Export',
-    path: '/settings/export',
-    exact: true,
-    component: Export
-  }
-];
