+++ common/v2/features/Dashboard/index.ts
@@ -1,2 +1,1 @@
-export { default } from './Dashboard';
-export { default as dashboardRoutes } from './routes';
+export { default as Dashboard } from './Dashboard';
