+++ common/v2/features/DevTools/index.ts
@@ -1,1 +1,1 @@
-export { default } from './DevTools';
+export { default as DevTools } from './DevTools';
