--- common/v2/features/DownloadApp/routes.ts
@@ -1,10 +1,0 @@
-import DownloadApp from './DownloadApp';
-
-export default [
-  {
-    name: 'Download Desktop App',
-    path: '/download-desktop-app',
-    exact: true,
-    component: DownloadApp
-  }
-];
