+++ common/v2/features/Layout/Layout.scss
@@ -1,6 +1,9 @@
 .Layout {
   min-width: 350px;
   background: #f6f8fa;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
 
   &-content {
     padding: 100px 0 50px 0;
@@ -10,6 +13,7 @@
         display: flex;
         align-items: center;
         justify-content: center;
+        flex: 1;
       }
     }
     @media (min-width: 850px) {
