+++ common/v2/features/Layout/index.ts
@@ -1,1 +1,1 @@
-export { default as Layout } from './Layout';
+export { default } from './Layout';
