+++ common/v2/features/NoAccounts/index.ts
@@ -1,3 +1,1 @@
-export { default } from './NoAccounts';
-export { default as noAccountsRoutes } from './routes';
-export { default as NoAccountAuth } from './NoAccountAuth';
+export { default as NoAccounts } from './NoAccounts';
