--- common/v2/features/NoAccounts/routes.ts
@@ -1,10 +1,0 @@
-import NoAccounts from './NoAccounts';
-
-export default [
-  {
-    name: 'No Accounts',
-    path: '/no-accounts',
-    exact: true,
-    component: NoAccounts
-  }
-];
