+++ common/v2/features/Nonce/index.ts
@@ -1,1 +1,1 @@
-export * from './nonceFunctions';
+export { getNonce } from './nonceFunctions';
