+++ common/v2/features/RequestAssets/RequestAssets.tsx
@@ -20,7 +20,7 @@ const truncate = (children: string) => {
   return [children.substring(0, 15), '…', children.substring(children.length - 10)].join('');
 };
 
-export function RequestAssets({ history }: RouteComponentProps<{}>) {
+function RequestAssets({ history }: RouteComponentProps<{}>) {
   return (
     <div className="RequestAssets">
       <ContentPanel
