+++ common/v2/features/RequestAssets/index.ts
@@ -1,2 +1,1 @@
 export { default as RequestAssets } from './RequestAssets';
-export { default as requestAssetsRoutes } from './routes';
