--- common/v2/features/RequestAssets/routes.tsx
@@ -1,10 +1,0 @@
-import { RequestAssets } from './RequestAssets';
-
-export default [
-  {
-    name: 'Request Assets',
-    path: '/request',
-    exact: true,
-    component: RequestAssets
-  }
-];
