+++ common/v2/features/SendAssets/index.ts
@@ -1,2 +1,1 @@
-export { default } from './SendAssets';
-export { default as sendAssetsRoutes } from './routes';
+export { SendAssets } from './SendAssets';
