--- common/v2/features/SendAssets/routes.ts
@@ -1,10 +1,0 @@
-import SendAssets from './SendAssets';
-
-export default [
-  {
-    name: 'Send Assets',
-    path: '/send',
-    exact: true,
-    component: SendAssets
-  }
-];
