+++ common/v2/features/index.ts
@@ -1,20 +1,10 @@
-export * from './BuyAndExchange';
-export * from './CreateWallet';
-export * from './Dashboard';
-export { Home } from './Home';
-export { addAccountRoutes } from './AddAccount';
-export * from './Dashboard/Settings';
-export * from './DownloadApp';
-export * from './NoAccounts';
-export * from './constants';
-export * from './DevTools';
-export * from './Drawer';
-export { AccountScreen, UnlockScreen } from './Drawer';
-export * from './Layout';
 export * from './Wallets';
-export * from './constants';
-export * from './ScreenLock';
-export * from './SendAssets';
-export * from './Gas';
-export * from './RequestAssets';
-export * from './Nonce';
+export { COLORS, BREAK_POINTS, GITHUB_RELEASE_NOTES_URL } from './constants';
+export { ScreenLockLocking } from './ScreenLock';
+export { fetchGasPriceEstimates, getDefaultEstimates } from './Gas';
+export { getNonce } from './Nonce';
+export { Drawer, AccountScreen, UnlockScreen, SelectLanguage } from './Drawer';
+export { DevTools } from './DevTools';
+export { Home } from './Home';
+export { default as Layout } from './Layout';
+export { default as appRoutes } from './routes';
