--- common/v2/features/registry.json
@@ -1,42 +1,0 @@
-[
-  {
-    "name": "Buy and Exchange",
-    "key": "buyAndExchange"
-  },
-  {
-    "name": "Create Wallet",
-    "key": "createWallet"
-  },
-  {
-    "name": "Dashboard",
-    "key": "dashboard"
-  },
-  {
-    "name": "Add Account",
-    "key": "addAccount"
-  },
-  {
-    "name": "Download App",
-    "key": "downloadApp"
-  },
-  {
-    "name": "No Accounts",
-    "key": "noAccounts"
-  },
-  {
-    "name": "Sceen Lock",
-    "key": "screenLock"
-  },
-  {
-    "name": "Settings",
-    "key": "settings"
-  },
-  {
-    "name": "Send Assets",
-    "key": "sendAssets"
-  },
-  {
-    "name": "Request Assets",
-    "key": "requestAssets"
-  }
-]
