+++ common/v2/index.ts
@@ -1,5 +1,4 @@
 export * from './features';
-export * from './routing';
 export * from './services';
 export * from './libs';
 export * from './config';
